+++ define/testdata/reference_graph/operation_graph_many_actions_integration/binding_hole_fans_out_to_multiple_fragments_and_multiple_callee_bindings/expected_trace/local/my_domain_com/my_lib/test/__init__.py
@@ -41,6 +41,7 @@ class TestExecution:
         )
         self.execution_position_gateway__action_middle: local.my_domain_com.my_lib.middle.MiddleExecution
         self.destruction_position_position_gateway__action_middle__position_trigger_pos: literal.Position
+        self.join_for_destroy_position_gateway = self.scheduler.create_join(3)
 
     def on_action_parent_occupied(self):
         self.create_position_gateway()
@@ -59,6 +60,12 @@ class TestExecution:
             self.scheduler,
             self.trace_execution,
             "middle",
+        )
+        self.execution_position_gateway__action_middle.guarantees.action_child_a__position_trigger_pos.consumers.append(
+            self.destroy_position_gateway
+        )
+        self.execution_position_gateway__action_middle.guarantees.action_child_b__position_trigger_pos.consumers.append(
+            self.destroy_position_gateway
         )
         self.scheduler.continue_with(
             self.create_position_gateway__action_middle__position_trigger_pos,
@@ -87,6 +94,11 @@ class TestExecution:
             "gateway::/middle::trigger_pos",
             1,
         )
+        self.destroy_position_gateway()
+
+    def destroy_position_gateway(self):
+        if not self.join_for_destroy_position_gateway.arrive():
+            return
         self.local_position_gateway.destroy_particle()
         self.scheduler.destroy_completed(
             self.trace_execution,
